requirements.txt
google-generativeai
pyyaml
asyncio